pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=1.22.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
